commit test cases & delete commit bug (#361)
* improve commit test cases

* improve commit test cases

* added Code & Environment config test cases

+++ backend/src/main/java/ai/verta/modeldb/versioning/CommitDAORdbImpl.java
@@ -2,6 +2,7 @@ package ai.verta.modeldb.versioning;
 
 import ai.verta.modeldb.App;
 import ai.verta.modeldb.ModelDBException;
+import ai.verta.modeldb.dto.CommitPaginationDTO;
 import ai.verta.modeldb.entities.versioning.CommitEntity;
 import ai.verta.modeldb.entities.versioning.RepositoryEntity;
 import ai.verta.modeldb.utils.ModelDBHibernateUtil;
@@ -55,7 +56,7 @@ public class CommitDAORdbImpl implements CommitDAO {
   }
 
   @Override
-  public List<CommitEntity> fetchCommitEntityList(
+  public CommitPaginationDTO fetchCommitEntityList(
       Session session, ListCommitsRequest request, Long repoId) throws ModelDBException {
     StringBuilder commitQueryBuilder =
         new StringBuilder(
@@ -95,7 +96,16 @@ public class CommitDAORdbImpl implements CommitDAO {
       commitEntityQuery.setFirstResult(startPosition);
       commitEntityQuery.setMaxResults(pageLimit);
     }
-    return commitEntityQuery.list();
+
+    Query countQuery = session.createQuery(commitQueryBuilder.toString());
+    countQuery.setParameter("repoId", repoId);
+    // TODO: improve query into count query
+    Long totalRecords = (long) countQuery.list().size();
+
+    CommitPaginationDTO commitPaginationDTO = new CommitPaginationDTO();
+    commitPaginationDTO.setCommitEntities(commitEntityQuery.list());
+    commitPaginationDTO.setTotalRecords(totalRecords);
+    return commitPaginationDTO;
   }
 
   @Override
@@ -104,11 +114,16 @@ public class CommitDAORdbImpl implements CommitDAO {
     try (Session session = ModelDBHibernateUtil.getSessionFactory().openSession()) {
       RepositoryEntity repository = getRepository.apply(session);
 
-      List<CommitEntity> commitEntities =
+      CommitPaginationDTO commitPaginationDTO =
           fetchCommitEntityList(session, request, repository.getId());
       List<Commit> commits =
-          commitEntities.stream().map(CommitEntity::toCommitProto).collect(Collectors.toList());
-      return ListCommitsRequest.Response.newBuilder().addAllCommits(commits).build();
+          commitPaginationDTO.getCommitEntities().stream()
+              .map(CommitEntity::toCommitProto)
+              .collect(Collectors.toList());
+      return ListCommitsRequest.Response.newBuilder()
+          .addAllCommits(commits)
+          .setTotalRecords(commitPaginationDTO.getTotalRecords())
+          .build();
     }
   }
 
@@ -182,13 +197,14 @@ public class CommitDAORdbImpl implements CommitDAO {
 
   @Override
   public DeleteCommitRequest.Response deleteCommit(
-      String commitHash, RepositoryFunction getRepository) throws ModelDBException {
+      DeleteCommitRequest request, RepositoryDAO repositoryDAO) throws ModelDBException {
     try (Session session = ModelDBHibernateUtil.getSessionFactory().openSession()) {
       session.beginTransaction();
-      RepositoryEntity repositoryEntity = getRepository.apply(session);
+      RepositoryEntity repositoryEntity =
+          repositoryDAO.getRepositoryById(session, request.getRepositoryId());
       boolean exists =
           VersioningUtils.commitRepositoryMappingExists(
-              session, commitHash, repositoryEntity.getId());
+              session, request.getCommitSha(), repositoryEntity.getId());
       if (!exists) {
         throw new ModelDBException(
             "Commit_hash and repository_id mapping not found", Code.NOT_FOUND);
@@ -199,8 +215,17 @@ public class CommitDAORdbImpl implements CommitDAO {
               "From "
                   + CommitEntity.class.getSimpleName()
                   + " c WHERE c.commit_hash = :commitHash");
-      deleteQuery.setParameter("commitHash", commitHash);
+      deleteQuery.setParameter("commitHash", request.getCommitSha());
       CommitEntity commitEntity = (CommitEntity) deleteQuery.uniqueResult();
+
+      if (!commitEntity.getChild_commits().isEmpty()) {
+        throw new ModelDBException(
+            "Commit has the child, please delete child commit first", Code.FAILED_PRECONDITION);
+      }
+
+      // delete associated branch
+      repositoryDAO.deleteBranchByCommit(repositoryEntity.getId(), request.getCommitSha());
+
       if (commitEntity.getRepository().size() == 1) {
         session.delete(commitEntity);
       } else {
